external/network-device-mcp-server/requirements.txt
# FastAPI Requirements - Python 3.13 Compatible

# Core FastAPI - Latest versions
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Pydantic - v2.10+ supports Python 3.13
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.2.0

# Database - SQLite only
sqlalchemy>=2.0.36
aiosqlite>=0.20.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1

# HTTP Client
httpx>=0.28.0

# Development & Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0

# Logging
loguru>=0.7.2